# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BilibiliComment Admin API
  version: 0.1.0
  description: 单管理员任务管理、评论观测与系统状态 API。所有 bigint ID 均按字符串传输。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/task-capabilities:
    get:
      tags: [Tasks]
      operationId: getTaskCapabilities
      responses:
        '200':
          description: 可创建的任务组合
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/TaskCapability'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/sources/resolve:
    post:
      tags: [Tasks]
      operationId: resolveSource
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResolveSourceRequest'}
      responses:
        '200':
          description: 规范化后的来源
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SourceResolution'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/tasks/summary:
    get:
      tags: [Tasks]
      operationId: getDashboardSummary
      responses:
        '200':
          description: 任务页摘要
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DashboardSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      parameters:
        - {name: query, in: query, schema: {type: string, maxLength: 255}}
        - {name: sourceType, in: query, schema: {$ref: '#/components/schemas/SourceType'}}
        - {name: collectionMode, in: query, schema: {$ref: '#/components/schemas/CollectionMode'}}
        - {name: runtimeState, in: query, schema: {$ref: '#/components/schemas/RuntimeState'}}
        - {name: health, in: query, schema: {$ref: '#/components/schemas/HealthStatus'}}
        - {name: scope, in: query, schema: {$ref: '#/components/schemas/TaskScope'}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 任务分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/Unprocessable'}
    post:
      tags: [Tasks]
      operationId: createTask
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTaskRequest'}
      responses:
        '201':
          description: 已创建；相同幂等请求会返回原任务
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskDetail'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/tasks/{taskId}:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Tasks]
      operationId: getTask
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tasks/{taskId}/discovered-tasks:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Tasks]
      operationId: listDiscoveredTasks
      parameters:
        - {name: query, in: query, schema: {type: string, maxLength: 255}}
        - {name: sourceType, in: query, schema: {type: string, enum: [VIDEO, DYNAMIC]}}
        - {name: runtimeState, in: query, schema: {$ref: '#/components/schemas/RuntimeState'}}
        - {name: health, in: query, schema: {$ref: '#/components/schemas/HealthStatus'}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 该 UP 主监控任务发现的内容及采集状态
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DiscoveredTaskPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/tasks/{taskId}/pause:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    post:
      tags: [Tasks]
      operationId: pauseTask
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 暂停后的任务
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/tasks/{taskId}/resume:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    post:
      tags: [Tasks]
      operationId: resumeTask
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 恢复后的任务
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/tasks/{taskId}/run:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    post:
      tags: [Executions]
      operationId: runTask
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 已排队的执行
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskExecution'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
...
    Cursor:
      name: cursor
      in: query
      description: 服务端返回的不透明分页边界；客户端必须原样回传，不得解析或自行构造。
      schema: {type: string, minLength: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
  responses:
    BadRequest:
      description: 请求格式错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unauthorized:
      description: 未登录或凭据错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Forbidden:
      description: CSRF 失效或来源域名被禁止
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    NotFound:
      description: 资源不存在
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Conflict:
      description: 重复任务、活动执行或幂等冲突
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
    Unprocessable:
      description: 参数验证失败
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    RateLimited:
      description: 请求过于频繁
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}}
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
    Unavailable:
      description: 数据库或上游暂不可用
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
  schemas:
    PositiveId: {type: string, pattern: '^[1-9][0-9]*$'}
...
    SourceType: {type: string, enum: [VIDEO, DYNAMIC, CREATOR]}
    TaskKind: {type: string, enum: [CONTENT_COMMENTS, CREATOR_WATCH]}
    CollectionMode: {type: string, enum: [FOLLOW_ONLY, BACKFILL_ONLY]}
    DesiredState: {type: string, enum: [ACTIVE, PAUSED]}
...
    RuntimeState: {type: string, enum: [IDLE, QUEUED, RUNNING, RETRY_WAIT, ERROR, COMPLETED]}
    HealthStatus: {type: string, enum: [UNKNOWN, HEALTHY, DEGRADED, ERROR]}
...
    TaskCapability:
      type: object
      required: [kind, sourceTypes, collectionModes]
      properties:
        kind: {$ref: '#/components/schemas/TaskKind'}
        sourceTypes: {type: array, items: {$ref: '#/components/schemas/SourceType'}}
        collectionModes: {type: array, items: {$ref: '#/components/schemas/CollectionMode'}}
    ResolveSourceRequest:
      type: object
      required: [sourceType, input]
      properties:
        sourceType: {$ref: '#/components/schemas/SourceType'}
        input: {type: string, minLength: 1, maxLength: 2048}
    TaskSource:
      type: object
      required: [type, id, title]
      properties:
        type: {$ref: '#/components/schemas/SourceType'}
        id: {type: string}
        title: {type: string}
...
    TaskScope:
      type: string
      enum: [PRIMARY, MANAGED, ALL]
      default: PRIMARY
...
    SourceResolution:
      type: object
      required: [source, accessible, message]
      properties:
        source: {$ref: '#/components/schemas/TaskSource'}
        accessible: {type: boolean}
        message: {type: [string, 'null']}
...
    TaskDetail:
      allOf:
        - {$ref: '#/components/schemas/TaskSummary'}
        - type: object
          required: [remark, createdAt, watchedContentTypes]
          properties:
            remark: {type: [string, 'null']}
            createdAt: {type: string, format: date-time}
            watchedContentTypes:
              type: array
              items: {type: string, enum: [VIDEO, DYNAMIC]}
...
    CreateTaskRequest:
      type: object
      required: [name, kind, source, collectionMode, credentialProfileId, schedule, desiredState, startNow]
      properties:
        name: {type: string, minLength: 1, maxLength: 255}
        kind: {$ref: '#/components/schemas/TaskKind'}
        source: {$ref: '#/components/schemas/TaskSourceInput'}
        collectionMode: {$ref: '#/components/schemas/CollectionMode'}
        credentialProfileId: {$ref: '#/components/schemas/PositiveId'}
        schedule: {$ref: '#/components/schemas/TaskScheduleInput'}
        watchedContentTypes:
          type: array
          uniqueItems: true
          items: {type: string, enum: [VIDEO, DYNAMIC]}
        desiredState: {$ref: '#/components/schemas/DesiredState'}
        startNow: {type: boolean}
        remark: {type: [string, 'null'], maxLength: 2000}
    TaskExecution:
      type: object
      required: [id, taskId, trigger, status, phase, pagesFetched, commentsDiscovered, commentsInserted, duplicatesSkipped, retryCount, cursor, errorSummary, startedAt, finishedAt, lastHeartbeatAt]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        taskId: {$ref: '#/components/schemas/PositiveId'}
        trigger: {type: string, enum: [SCHEDULED, MANUAL, DISCOVERED]}
        status: {$ref: '#/components/schemas/ExecutionStatus'}
        phase: {$ref: '#/components/schemas/ExecutionPhase'}
        pagesFetched: {type: integer, minimum: 0}
        commentsDiscovered: {type: integer, minimum: 0}
        commentsInserted: {type: integer, minimum: 0}
        duplicatesSkipped: {type: integer, minimum: 0}
        retryCount: {type: integer, minimum: 0}
        cursor: {type: [string, 'null']}
        errorSummary: {type: [string, 'null']}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        lastHeartbeatAt: {type: [string, 'null'], format: date-time}
...
    DashboardSummary:
      type: object
      required: [activeTasks, primaryActiveTasks, managedActiveTasks, runningTasks, retryWaitingTasks, commentsInserted24h, recentFailures]
      properties:
        activeTasks: {type: integer, minimum: 0}
        primaryActiveTasks: {type: integer, minimum: 0}
        managedActiveTasks: {type: integer, minimum: 0}
        runningTasks: {type: integer, minimum: 0}
        retryWaitingTasks: {type: integer, minimum: 0}
        commentsInserted24h: {type: integer, minimum: 0}
        recentFailures: {type: integer, minimum: 0}
...
    TaskPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/TaskSummary'}}
        nextCursor:
          type: [string, 'null']
          description: 下一页的不透明任务边界；为空表示没有下一页。
        total: {type: integer, minimum: 0}
    DiscoveredTaskPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DiscoveredTaskItem'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
...